Moved the App into its own component
